requirements.txt
flask
pytesseract
python-docx
fpdf
pillow
